Split the specs to match the source files they cover
The specs now line up one to one with the files under test, the way
LicensePluginAndroidSpec always did:

  ProjectJvm      LicensePluginJvmSpec
  ProjectAndroid  LicensePluginAndroidSpec
  ProjectKmp      LicensePluginKmpSpec

LicensePluginJavaSpec is renamed rather than copied, so history follows, and it
covers Kotlin JVM as well as Java, which is the same reason ProjectJava became
ProjectJvm.

Only the multiplatform project test moves. The kotlin multiplatform dedup test
stays put: it is a plain java-library project that happens to depend on KMP
artifacts, not a multiplatform project.

classpathString was added to the Java spec only to give that one test the Kotlin
Gradle Plugin, so it moves with it rather than being left behind unused.

35 + 1 tests, matching the 36 the single spec had.
